pyproject.toml
[project]
name = "pip-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paramiko-ng>=2.9.0",
    "pyqt6>=6.9.0",
]
